Remove quantized_vector field from VectorRecord proto and fix compilation errors
- Removed quantized_vector field from proto definition
- Fixed all references in test files and implementations
- Updated documentation to reflect accurate feature status
- Reorganized roadmap documentation with correct completion percentages

+++ src/storage/engines/core/formats/columnar/columnar_query_engine/unified_reader.rs
@@ -850,7 +850,7 @@ impl UnifiedParquetReader {
                 id,
                 vector,
                 metadata,
-                version: Some(version),
+                version: Some(version as u32),
                 timestamp,
                 ..Default::default()
             });
